# spec/openapi.yaml
openapi: 3.1.0
info:
  title: ArchForge API
  version: 0.1.0
  summary: Stub contract for ArchForge admin and C-end APIs.
  description: |
    Two servers, two clients:

    - **Admin** (`server-admin`, port 8080) — consumed by ArchForgeAdmin.
      Success envelope `{code, message, data}`. See `specs/api-response.md`.
    - **Web** (`server-web`, port 8081) — consumed by ArchForgeWeb.
      Errors use RFC 9457 ProblemDetail.

    This file is a **stub**. Paths listed here are the live surface we are
    willing to document. Do **not** reintroduce deleted `/system/menu` or
    `/system/role`. Leftover `/system/dict` is a migration backlog item
    (see `specs/api-path.md`), not a target prefix.

    When an endpoint changes, update this file in the same change.
servers:
  - url: http://localhost:8080
    description: server-admin (ArchForgeAdmin)
  - url: http://localhost:8081
    description: server-web (ArchForgeWeb)
...
security:
  - saToken: []
paths:
...
  /quartz/update/{id}:
    put:
      tags: [admin-legacy]
      summary: Update scheduled job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          $ref: "#/components/responses/AdminOk"
  /quartz/{id}:
    delete:
      tags: [admin-legacy]
      summary: Delete scheduled job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          $ref: "#/components/responses/AdminOk"
  /quartz/pause/{id}:
    post:
      tags: [admin-legacy]
      summary: Pause scheduled job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          $ref: "#/components/responses/AdminOk"
  /quartz/resume/{id}:
    post:
      tags: [admin-legacy]
      summary: Resume scheduled job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          $ref: "#/components/responses/AdminOk"
  /quartz/run/{id}:
    post:
      tags: [admin-legacy]
      summary: Run scheduled job once
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          $ref: "#/components/responses/AdminOk"
...
components:
  securitySchemes:
    saToken:
      type: http
      scheme: bearer
      description: "sa-token value. Header Authorization: Bearer <token>."
  responses:
    AdminOk:
      description: Admin envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AdminEnvelope"
...
  schemas:
    AdminEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 0 means success
        message:
          type: string
        data: {}
...